requirements.txt
Flask-SQLAlchemy==3.0.3
Flask==2.3.2
Jinja2==2.11.3
SQLAlchemy==2.0.15
Werkzeug==2.3.4
aiohttp
python-dateutil==2.8.2
python-telegram-bot==13.13
requests
solana